Disable Page Fusion feature, does not work in VirtualBox 7.2
Attempting to save a VM with Page Fusion on or off results in error in VirtualBox 7.2.
"Page fusion is only supported on 64-bit hosts". Tested on 64-bit Linux 6.12 and Windows 11.
Disable the feature for now, hopefully this prevents the error from getting triggered when
saving old VMs. I had no way to test this, but at least disabling this won't trigger
the issue in new VMs.
Issue #363