Fix missing NVRAM file when manually checking Enable EFI in System settings
NVRAM file is not created for new VMs where EFI is not set by defalut
(I guess pretty much anything except Windows 11 and macOS).
When manually checking "Enable EFI" in System settings later, the file was not
created, and the VM settings became unavailable. I was trying to make the secure
boot flag working, but due to the issue with vboxwebsrv always reporting it
as enabled, and inability to tell whether it's disabled, I had to turn this off,
and carelessly also commented the code creating the NVRAM file.
This part is now restored. Fixes #369